backend/pyproject.toml
[project]
name = "mafia-bot"
version = "0.1.0"
description = "Mafia Telegram Bot — @MafGameUzBot"
requires-python = ">=3.12"
dependencies = [
    # Bot
    "aiogram>=3.13.0",
    "redis>=5.1.0",

    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "websockets>=13.1",

    # ORM (asyncpg for postgres, aiosqlite for local dev)
    "tortoise-orm[asyncpg]>=0.21.6",
    "aiosqlite>=0.20.0",
    "aerich>=0.7.2",

    # Settings
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",

    # i18n
    "fluent.runtime>=0.4.0",

    # Logging & monitoring
    "loguru>=0.7.2",
    "sentry-sdk[fastapi]>=2.15.0",

    # Scheduler
    "apscheduler>=3.10.4",

    # Auth
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.9.0",
    "python-multipart>=0.0.12",

    # Utils
    "httpx>=0.27.0",
    "orjson>=3.10.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "pre-commit>=3.8.0",
    "types-redis>=4.6.0",
]

[tool.aerich]
tortoise_orm = "app.db.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./app"

# === Ruff ===
[tool.ruff]
line-length = 100
target-version = "py312"
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
    "RUF", # ruff-specific
    "ASYNC", # flake8-async
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # function calls in argument defaults (FastAPI Depends)
    "B904",  # raise from in except
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG"]
"app/db/models/*.py" = ["RUF012"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# === MyPy ===
[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
no_strict_optional = true
warn_unused_ignores = false
disallow_untyped_defs = false
plugins = ["pydantic.mypy"]

# === Pytest ===
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short --strict-markers"
